requirements.txt
numpy==1.24.4
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
matplotlib==3.7.3
tqdm==4.66.1